requirements.txt
kafka-python
clickhouse-connect